requirements.txt
# Core App
streamlit>=1.40.0
python-dotenv==1.0.1

# FastAPI Backend
fastapi>=0.110.0
uvicorn>=0.29.0

# LLM (Gemini)
google-generativeai>=0.7.0

# Vector Database
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# PDF Processing
PyMuPDF>=1.24.3

# Image Processing & OCR
pillow>=10.4.0
pytesseract==0.3.10

# Data Handling
pandas>=2.2.0
numpy>=2.0.0

# ML / Analytics
scikit-learn>=1.5.0

# Utilities
tqdm>=4.66.1

# New Features
fpdf2>=2.7.0
httpx>=0.27.0
python-multipart>=0.0.9
